Restructured say and owo commands and imported them to the Searches module
