+++ cogs/basic.py
@@ -1,5 +1,4 @@
 from discord.ext import commands
-from utils import text_to_owo
 
 class Basic(commands.Cog):
     def __init__(self, bot):
@@ -10,20 +9,7 @@ class Basic(commands.Cog):
         print(exception)
         await ctx.send("Please check !help for tips on how to use this command or ask an administrator.")
 
-    
-    @commands.command(description="Provide arguments", brief="The bot will repeat what you typed")
-    async def say(self, ctx, *args):
-        if len(args) > 0:
-            await ctx.send(" ".join(args))
-        else:
-            await ctx.send("It is not possible to send an empty message.")
-    
 
-    @commands.command(brief="Text-to-OwO")
-    async def owo(self, ctx):
-        await ctx.send(text_to_owo(ctx.message.content[5:]))
-
-    
     @commands.command(brief="Creates an invite-link for the current channel, which will last for 1 day")
     @commands.guild_only()
     async def invite(self, ctx):
